Added an interrupt request code for a memory manager interrupt.This allows the memory manager to generate  a software interrupt when an a memory related error occurrs. Added an interrupt routine to handle MMU IRQs. Modified the MMU as follows: 	Replaced magic numbers with constants. 	When an access violation or out-of-memory access error occurs the MMU generates a software interrupt rather than trapping to the kernel.
Modified the shellLoad command to parse the text in the program input text area and pass it to the kernel. Previously, shellLoad wrote the source to main memory before calling the krnNewProcess command. This caused a potential issue when the load command was called when there was no memory available. It would simply overrwrite main memory...very bad.